feat: make output show more info about env
